fix: Planned transactions page shows transactions correctly on tablet
