+++ src/ducks/future/styles.styl
@@ -4,6 +4,9 @@
 .List--mobile
     margin-top 6.5rem
 
+.List--tablet
+    margin-top 8.5rem
+
 .Title--mobile
   margin-top -3.25rem
   position relative
